+++ components/Column/index.tsx
@@ -9,9 +9,9 @@ import {
   getUnits,
 } from '@shallot-ui/core'
 
-type BoxShallot = ShallotProp
+type ColumnShallot = ShallotProp
 
-export type BoxStyleProps = {
+export type ColumnStyleProps = {
   backgroundColor?: AllColorShades
   borderColor?: AllColorShades
   radius?: keyof DefaultTheme['radii']
@@ -28,11 +28,12 @@ export type BoxStyleProps = {
   minUnitWidth?: number
 }
 
-export type BoxProps<T = {}> = T & BoxStyleProps & { shallot?: BoxShallot }
+export type ColumnProps<T = {}> = T &
+  ColumnStyleProps & { shallot?: ColumnShallot }
 
-export const withBoxStyleProps =
-  <T,>(BoxComponent: ComponentType<T & { shallot?: BoxShallot }>) =>
-  (props: BoxProps<T>) => {
+export const withColumnStyleProps =
+  <T,>(ColumnComponent: ComponentType<T & { shallot?: ColumnShallot }>) =>
+  (props: ColumnProps<T>) => {
     const {
       backgroundColor,
       borderColor,
@@ -52,8 +53,9 @@ export const withBoxStyleProps =
       ...nonStyleProps
     } = props
 
-    let boxShallot: BoxShallot = {
+    let boxShallot: ColumnShallot = {
       display: 'flex',
+      flexDirection: 'column',
       ...(radius && { borderRadius: getRadius(radius) }),
       ...(backgroundColor && {
         backgroundColor: getColorShade(backgroundColor),
@@ -74,5 +76,5 @@ export const withBoxStyleProps =
 
     boxShallot = applyStyles(boxShallot, shallot)
 
-    return <BoxComponent {...(nonStyleProps as T)} shallot={boxShallot} />
+    return <ColumnComponent {...(nonStyleProps as T)} shallot={boxShallot} />
   }