+++ platforms/web/components/Box/index.tsx
@@ -2,5 +2,5 @@ import styled from 'styled-components'
 import { ShallotProp, getStyle } from '@shallot-ui/core'
 import { withBoxStyleProps } from '@shallot-ui/box'
 
-const RawBox = styled.div<{ shallot?: ShallotProp }>(getStyle)
-export const Box = withBoxStyleProps(RawBox)
+const Base = styled.div<{ shallot?: ShallotProp }>(getStyle)
+export const Box = withBoxStyleProps(Base)