+++ platforms/web/components/Row/index.ts
@@ -1,0 +1,6 @@
+import styled from 'styled-components'
+import { ShallotProp, getStyle } from '@shallot-ui/core'
+import { withRowStyleProps } from '@shallot-ui/row'
+
+const Base = styled.div<{ shallot?: ShallotProp }>(getStyle)
+export const Row = withRowStyleProps(Base)