+++ platforms/web/components/content/index.ts
@@ -1,3 +1,0 @@
-export * from './AnchorLink'
-export * from './Tag'
-export * from './Text'