+++ components/Text/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @shallot-ui/text
 
+## 0.4.0
+
+### Minor Changes
+
+- More style props
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/box@0.4.0
+  - @shallot-ui/theme@0.27.0
+  - @shallot-ui/core@5.0.0
+
+## 0.3.0
+
+### Minor Changes
+
+- Adds some style props bacl
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/box@0.3.0
+  - @shallot-ui/theme@0.26.0
+  - @shallot-ui/core@4.0.0
+
 ## 0.2.0
 
 ### Minor Changes