+++ packages/core/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @shallot-ui/core
 
+## 5.0.0
+
+### Minor Changes
+
+- More style props
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/theme@0.27.0
+
+## 4.0.0
+
+### Minor Changes
+
+- Adds some style props bacl
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/theme@0.26.0
+
 ## 3.0.0
 
 ### Minor Changes