+++ platforms/next/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @shallot-ui/next
 
+## 0.27.0
+
+### Minor Changes
+
+- More style props
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/web@0.41.0
+
+## 0.26.0
+
+### Minor Changes
+
+- Adds some style props bacl
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/web@0.40.0
+
 ## 0.25.0
 
 ### Minor Changes